+++ quake3-rs/qagamex86_64/src/game/g_session.rs
@@ -163,7 +163,7 @@ pub unsafe extern "C" fn G_WriteClientSessionData(mut client: *mut crate::g_loca
     );
     var = crate::src::qcommon::q_shared::va(
         b"session%i\x00" as *const u8 as *const libc::c_char as *mut libc::c_char,
-        client.offset_from(crate::src::game::g_main::level.clients) as libc::c_long as i32,
+        client.offset_from(crate::src::game::g_main::level.clients) as isize as i32,
     );
     crate::src::game::g_syscalls::trap_Cvar_Set(var, s);
 }
@@ -184,7 +184,7 @@ pub unsafe extern "C" fn G_ReadSessionData(mut client: *mut crate::g_local_h::gc
     let mut sessionTeam: i32 = 0;
     var = crate::src::qcommon::q_shared::va(
         b"session%i\x00" as *const u8 as *const libc::c_char as *mut libc::c_char,
-        client.offset_from(crate::src::game::g_main::level.clients) as libc::c_long as i32,
+        client.offset_from(crate::src::game::g_main::level.clients) as isize as i32,
     );
     crate::src::game::g_syscalls::trap_Cvar_VariableStringBuffer(
         var,
@@ -254,7 +254,7 @@ pub unsafe extern "C" fn G_InitSessionData(
                 &mut *crate::src::game::g_main::g_entities
                     .as_mut_ptr()
                     .offset(client.offset_from(crate::src::game::g_main::level.clients)
-                        as libc::c_long as isize) as *mut _
+                        as isize) as *mut _
                     as *mut crate::g_local_h::gentity_s,
                 value,
             );
