+++ quake3-rs/qagamex86_64/src/game/g_target.rs
@@ -489,7 +489,7 @@ pub unsafe extern "C" fn Use_Target_Print(
 ) {
     if !(*activator).client.is_null() && (*ent).spawnflags & 4 as i32 != 0 {
         crate::src::game::g_syscalls::trap_SendServerCommand(
-            activator.offset_from(crate::src::game::g_main::g_entities.as_mut_ptr()) as libc::c_long
+            activator.offset_from(crate::src::game::g_main::g_entities.as_mut_ptr()) as isize
                 as i32,
             crate::src::qcommon::q_shared::va(
                 b"cp \"%s\"\x00" as *const u8 as *const libc::c_char as *mut libc::c_char,
